added a lot of post functionality

+++ server.js
@@ -30,27 +30,64 @@ app.get("/", (req, res) => {
 		message: "Welcome to the Post Service of Learn Sphere!",
 	});
 });
-// ---------------------------POST CREATE METHOD-------------------------
+// ---------------------------POST CREATE METHOD------------------------- //need to add youtube URL, and title, possible change comments array to have comment posted by, comment id.
 app.post("/posts/create", async (req, res) => {
 	try {
-		console.log(req.body);
-		const postJson = {
-			postId: uuidv4(),
-			createdAt: new Date(),
-			comments: [],
-			postedBy: req.body.teacherId,
-			image: req.body.image,
-			description: req.body.description,
-		};
-		const response = await db
-			.collection("posts")
-			.doc(postJson.postId)
-			.set(postJson);
-		res.send(response);
+		const { username, role } = res.locals.user;
+		if(role === "admin" || role === "teacher"){
+			const image = req.body.image || "http://placehold.it/300x200";
+			const description = req.body.description || "No description";
+
+			const postJson = {
+				postId: uuidv4(),
+				createdAt: new Date(),
+				comments: [],
+				postedBy: req.body.userId,
+				image: image,
+				description: description,
+			};
+			const response = await db
+				.collection("posts")
+				.doc(postJson.postId)
+				.set(postJson);
+				
+			console.log("Post has been created!: \n", postJson);
+			res.send(response);
+		}
 	} catch (error) {
 		res.send(error);
 	}
 });
+
+// app.post("/posts/create", async (req, res) => {
+//     try {
+//         const { username, role } = res.locals.user;
+//         if (role === "admin" || role === "teacher") {
+//             const postJson = {
+//                 postId: uuidv4(),
+//                 createdAt: new Date(),
+//                 comments: {},
+//                 postedBy: req.body.userId,
+//                 image: req.body.image,
+//                 description: req.body.description,
+//                 lectureURL: req.body.lectureURL, // Add lectureURL field
+//                 title: req.body.title
+//             };
+
+
+//             const response = await db
+//                 .collection("posts")
+//                 .doc(postJson.postId)
+//                 .set(postJson);
+
+//             console.log("Post has been created!: \n", postJson);
+//             res.send(response);
+//         }
+//     } catch (error) {
+//         res.send(error);
+//     }
+// });
+
 
 // ----------------------------GET A POST USING POST ID METHOD------------------
 
@@ -75,6 +112,27 @@ app.get("/posts/:postId", async (req, res) => {
 	}
 });
 
+//----------------------------GET A POST USING USER ID------------------
+
+app.get("/posts/user/:userId", async (req, res) => {
+    try {
+        const { userId } = req.params;
+
+        const querySnapshot = await db.collection("posts").where("postedBy", "==", userId).get();
+
+        const posts = [];
+
+        querySnapshot.forEach((doc) => {
+            posts.push(doc.data());
+        });
+
+        res.json(posts);
+    } catch (error) {
+        console.error("Error fetching posts by userId:", error);
+        res.status(500).json({ message: "Internal Server Error" });
+    }
+});
+
 // ---------------------------------GETTING ALL POSTS THAT EXIST------------------------
 
 app.get("/posts", async (req, res) => {
@@ -96,9 +154,28 @@ app.get("/posts", async (req, res) => {
 app.delete("/posts/delete/:postId", async (req, res) => {
 	try {
 		const { postId } = req.params;
-		console.log("Deleting post that his this postId: ", postId);
-		const response = await db.collection("posts").doc(postId).delete();
-		res.send(response);
+		const { username, role } = res.locals.user;
+
+		const query_return = await db
+			.collection("posts")
+			.where("postId", "==", postId)
+			.get();
+		
+		if (!query_return.empty) {
+			const doc = query_return.docs[0];
+			const postData = doc.data();
+			if(role === "admin" || username === postData.postedBy){
+				console.log("Deleting post with postId", postId, "deleted by", username, role);
+				const response = await db.collection("posts").doc(postId).delete();
+				res.send(response);
+			} else{
+				res.send("Permission error");
+			}
+		} else {
+			res.status(404).send("Post not found");
+		}
+		
+		
 	} catch (error) {
 		res.send(error);
 	}
@@ -108,25 +185,110 @@ app.delete("/posts/delete/:postId", async (req, res) => {
 app.put("/posts/update/:postId", async (req, res) => {
 	try {
 		const { postId } = req.params;
+		const { username, role } = res.locals.user;
+
 		const updateFields = req.body;
 
 		const postRef = db.collection("posts").doc(postId);
 		const postDoc = await postRef.get();
-
 		if (!postDoc.exists) {
 			return res.status(404).send("Post not found");
 		}
 
-		const updatedPostData = { ...postDoc.data(), ...updateFields };
+		const query_return = await db
+			.collection("posts")
+			.where("postId", "==", postId)
+			.get();
+		if (!query_return.empty) {
+			const doc = query_return.docs[0];
+			const postData = doc.data();
+			if(role === admin || username === postData.postedBy){ //permissions
 
-		await postRef.update(updatedPostData);
+				const updatedPostData = { ...postDoc.data(), ...updateFields };
+				await postRef.update(updatedPostData);
+				console.log("Post",postId,"has been updated", postData);	
+				res.status(200).send("Post updated successfully");
 
-		res.status(200).send("Post updated successfully");
+			} else{
+				res.send("Permission error");
+			}
+		}
+
+		
 	} catch (error) {
 		console.error("Error updating post:", error);
 		res.status(500).send("Internal Server Error");
 	}
 });
+
+//----------------------------------------COMMENT ON A POST--------------------------------------
+app.post("/posts/:postId/addcomment", async (req, res) => {
+    try {
+        const { postId } = req.params;
+        const { comment } = req.body;
+//	
+//             req.body.comments.forEach(comment => {
+//             {
+//                     author: req.body.userId,
+//                     id: uuidv4(),
+//                     comment: req.body.comment
+//             }
+
+        const postRef = db.collection("posts").doc(postId);
+        const postDoc = await postRef.get();
+
+        if (!postDoc.exists) {
+            return res.status(404).json({ message: "Post not found" });
+        }
+
+        const updatedComments = [...postDoc.data().comments, comment];
+        await postRef.update({ comments: updatedComments });
+
+		console.log("Comment added on",postId, comment);
+        return res.status(200).json({ message: "Comment added successfully" });
+    } catch (error) {
+        console.error("Error adding comment:", error);
+        return res.status(500).json({ message: "Internal Server Error" });
+    }
+});
+
+//----------------------------------------DELETE A COMMENT--------------------------------------
+app.delete("/posts/:postId/comments/:commentId/delete", async (req, res) => {
+    try {
+        const { postId, commentId } = req.params;
+        const { username, role } = res.locals.user;
+
+        const postRef = db.collection("posts").doc(postId);
+        const postDoc = await postRef.get();
+
+        if (!postDoc.exists) {
+            return res.status(404).json({ message: "Post not found" });
+        }
+
+        const comments = postDoc.data().comments;
+
+        if (commentId === -1) {
+            return res.status(404).json({ message: "Comment not found" });
+        }
+
+        const comment = comments[commentId];
+
+        if (role !== "admin") {
+            return res.status(403).json({ message: "Unauthorized to delete this comment" });
+        }
+
+        comments.splice(commentId, 1);
+
+        await postRef.update({ comments });
+
+		console.log("Comment deleted on",postId, comment, "commendId:", commentId);
+        return res.status(200).json({ message: "Comment deleted successfully" });
+    } catch (error) {
+        console.error("Error deleting comment:", error);
+        return res.status(500).json({ message: "Internal Server Error" });
+    }
+});
+
 
 app.listen(PORT, () => {
 	console.log(`Post Service is listing on PORT ${PORT}...`);
